Fix decimal field

+++ backend/weight_tracker/models.py
@@ -10,8 +10,8 @@ class UserMetrics(models.Model):
         verbose_name_plural = "метрики"
 
     user = models.ForeignKey(User, on_delete=models.CASCADE, verbose_name="Пользователь", related_name="metrics")
-    weight = models.PositiveIntegerField(default=0, verbose_name="Вес (кг)")
-    waist_circumference = models.PositiveIntegerField(default=None, null=True, blank=True, verbose_name="Обхват талии")
+    weight = models.DecimalField(default=0, max_digits=2, decimal_places=2, verbose_name="Вес (кг)")
+    waist_circumference = models.DecimalField(default=None, null=True, blank=True, max_digits=2, decimal_places=2, verbose_name="Обхват талии")
     measurement_date = models.DateField(default=datetime.date.today, verbose_name="Дата измерения")
 
     def __str__(self):
@@ -28,7 +28,7 @@ class UserGoal(models.Model):
         GLOBAL = ("global", "Глобальная цель")
 
     user = models.ForeignKey(User, on_delete=models.CASCADE, verbose_name="Пользователь", related_name="goals")
-    weight_goal = models.PositiveIntegerField(default=0, verbose_name="Ожидаемый вес (кг)")
+    weight_goal = models.DecimalField(default=0, max_digits=2, decimal_places=2, verbose_name="Ожидаемый вес (кг)")
     goal_type = models.CharField(max_length=255, choices=GoalType.choices, verbose_name="Тип цели")
     target_date = models.DateField(
         default=datetime.date.today,
